[Remote suggestions] Reduce database memory use for low end devices
This CL changes the buffer size for leveldb database for remote
suggestions. This smaller buffer size is used only for devices with
memory pressure. The change saves up to 500 KB of memory.

Similarly as the previous buffer size of 512 KB, the new buffer size
of 128 KB was chosen arbitrarily (the largest size that still shows
considerable memory savings when tracing the process).

The CL also makes small changes to avoid empty calls to the database.

Bug: 716170

+++ components/ntp_snippets/remote/remote_suggestions_provider_impl.cc
@@ -1296,16 +1296,18 @@ void RemoteSuggestionsProviderImpl::ClearExpiredDismissedSuggestions() {
     }
     RemoveNullPointers(&content->dismissed);
 
-    // Delete the images.
-    database_->DeleteImages(GetSuggestionIDVector(to_delete));
-    // Delete the removed article suggestions from the DB.
-    database_->DeleteSnippets(GetSuggestionIDVector(to_delete));
-
     if (content->suggestions.empty() && content->dismissed.empty() &&
         category != articles_category_ &&
         !content->included_in_last_server_response) {
       categories_to_delete.push_back(category);
     }
+
+    if (to_delete.empty())
+      continue;
+    // Delete the images.
+    database_->DeleteImages(GetSuggestionIDVector(to_delete));
+    // Delete the removed article suggestions from the DB.
+    database_->DeleteSnippets(GetSuggestionIDVector(to_delete));
   }
 
   DeleteCategories(categories_to_delete);
